Fix stack support through an additional error type
Stack support was included but didn't work properly when outputting the
format. The intention was for `%+v` to be used when the stack should
show up, but this didn't work properly because `errors.Join` and
`fmt.Errorf` don't pass the `%+v` through when constructing the output.

This updates the library to include a `New` method that works the exact
same as `fmt.Errorf` and uses `fmt.Errorf` on the backend. It wraps the
output of that with an `errorValue` and `errorValue` properly handles
the formatting directive for the underlying stacks.

Stacks are also included by wrapping an error using `WithStack` rather
than using `errors.Join`. `errors.Join` would always create a newline
for the error and didn't have a way to hide the stack trace depending on
the formatting directive. We now just embed the stack trace with the
error and `errorValue` is what handles extracting the stacks.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>