Refactor errdefs with an API more similar to the standard library
This updates errdefs to have an API similar to the standard library. The
`New` method is the equivalent of `fmt.Errorf` and `Join` is the
equivalent of `errors.Join`.

These methods also support the proper output of stack traces by ensuring
the proper handling of the formatting for collapsible errors.

Similarly, the `stack` package has been updated to remove the
context-based helper and instead directly expose the function that
creates the stack error. The stack error is also renamed to `Error` and
exposed rather than left unexported.

The `stack` package can be used to directly create errors with stacks
when one isn't present. It is also possible to add multiple stacks to a
single error through `errdefs.Join` and `errdefs.ErrStack` to manually
create a stack error.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>